pyproject.toml
[project]
name = "medical-mcp"
version = "0.1.0"
description = "Chatbot MCP: weather, SERP web search, Wikipedia via FastMCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.27",
    "google-generativeai>=0.7.0",
    "streamlit>=1.36",
    "pydantic>=2.11.7",
    "wikipedia-api>=0.8.1",
]

